// src/internal/types/cc-mngt/cc-mntg-types.ts
/*
 * Copyright 2021 ALE International
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy of this
 * software and associated documentation files (the "Software"), to deal in the Software
 * without restriction, including without limitation the rights to use, copy, modify, merge,
 * publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons
 * to whom the Software is furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all copies or
 * substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING
 * BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
 * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
 * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */

import { PilotOperatingMode } from '../../../types/cc-mngt/calendar/pilot-operating-Mode';
import { ServiceState } from '../../../types/common/service-state';
import { PilotStatus } from '../../../types/telephony/call/ccd/pilot-status';
import { DayOfWeekJson } from '../common/common-types';

/** @internal */
export type PilotRuleJson = {
    ruleNumber: number;
    name?: string;
    active: boolean;
};

/** @internal */
export type PilotJson = {
    number?: string;
    name?: string;
    state?: ServiceState;
    detailedState?: PilotStatus;
    waitingTime?: number;
    saturation?: boolean;
    rules?: PilotRuleJson[];
    possibleTransfer?: boolean;
    supervisedTransfer?: boolean;
};

/** @internal */
export type PilotTransitionJson = {
    time: string;
    ruleNumber: number;
    mode: PilotOperatingMode;
};

/** @internal */
export type PilotTransitionsJson = {
    number: number;
    transition: PilotTransitionJson;
};

//------------------------------------------------------------
// NormalTransitions
//------------------------------------------------------------
/** @internal */
export type NormalTransitionsJson = {
    day: DayOfWeekJson;
    list?: PilotTransitionsJson[];
};

//------------------------------------------------------------
// ExceptionTransitions
//------------------------------------------------------------
/** @internal */
export type ExceptionTransitionsJson = {
    date: string;
    list?: PilotTransitionsJson[];
};

//------------------------------------------------------------
// O2GExceptionCalendar
//------------------------------------------------------------
/** @internal */
export type ExceptionCalendarJson = {
    calendar?: ExceptionTransitionsJson[];
};

//------------------------------------------------------------
// O2GNormalCalendar
//------------------------------------------------------------
/** @internal */
export type NormalCalendarJson = {
    calendar?: NormalTransitionsJson[];
};

//------------------------------------------------------------
// O2GNormalCalendar
//------------------------------------------------------------
/** @internal */
export type CalendarJson = {
    normalDays?: NormalCalendarJson;
    exceptionDays?: ExceptionCalendarJson;
};